Make context switch code portable
Currently the code performing context switch resides in platform support
code even if most of it is rather generic and independent on actual
platform. The code is split into two parts.

The generic part is moved out into the platform independent part of the
kernel. This portion determines pointers to old and new thread and
process. It the fills context switch structure and requests context
switch.

The platform support part now contains only the heavy lifting part
of the code which actually performs the switch.

As a part of this refactor, context switch context has been collected
into one structure, instead of many free-standing variables. Just
a little housekeeping, no real impact on functionality. Some unused
variables have been removed from the context as they were either
redundant or entirely unused.

+++ include/cmrx/os/runtime.h
@@ -114,7 +114,7 @@ struct OS_thread_t {
  */
 struct OS_stack_t {
 	/** Thread stacks. */
-	unsigned long stacks[OS_STACKS][OS_STACK_DWORD];
+	uint32_t stacks[OS_STACKS][OS_STACK_DWORD];
 
 	/** Information about stack allocation. If n-th bit is set, 
 	 * then n-th stack is allocated. Otherwise it is available. */
